Add Jabber invisible mode
git-svn-id: file:///.data/svn_convert/svn@2031 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/api/contacts.cpp
@@ -1014,14 +1014,14 @@ Client::Client(Protocol *protocol, const char *cfg)
         do {
             QString sub_str = pswd.left(4);
             pswd.remove(0,5);
-            
+
             temp ^= sub_str.toUShort(0,16);
             new_pswd += tmp.setUnicodeCodes(&temp,1);
             temp = sub_str.toUShort(0,16);
         } while (pswd.length());
         setPassword(new_pswd);
     }
-    
+
     m_protocol = protocol;
     m_status = STATUS_OFFLINE;
     m_state  = Offline;
@@ -1087,15 +1087,15 @@ string Client::getConfig()
     if (pswd.length()) {
         QString new_passwd;
         unsigned short temp = 0x4345;
-        for (unsigned int i = 0; i < pswd.length(); i++) {
+        for (int i = 0; i < (int)(pswd.length()); i++) {
             temp ^= (pswd[i].unicode());
             new_passwd += QString::number(temp,16);
             if (i+1 != pswd.length())
                 new_passwd +=";";
         }
         setPassword(new_passwd);
-    }    
-    
+    }
+
     if (!getSavePassword()){
         setPassword(NULL);
     }
